moved views.py into subviews and renamed as home.py
