+++ src/nac/subviews/home.py
